Fat: Generate app icon as a glyph
In Nanos, the app icon was not displayed inside the app.
Thus, it was not generated as a glyph.
In Stax, the app is now displayed on the Welcome page of the app.
This commits enables the generation of the app icon glyphs, as soon as
the app is built.